multiple additions & improvements
### Additions
#### Saving & Loading
Userdata (workoutHistory, user profile...) is now correctly stored on pause and quit, and correctly loaded on Awake() via JSON (de-)serialization

#### Alert-Box
Added an Alert-Box; currently used to alert the user if he tries to change the View whilst working out

#### Calendar
The calendar can be found under "Progress". Calendar cells color themselves based on the completed workouts on each day.

Reward-Colors are currently bronze, silver and gold; the thresholds are declared in the GameManager [medThresh, highThresh].

#### Workout exit and pause/resume
Added the option to exit a workoutSession early and to pause/resume a workoutSession.

#### Added basic DailyChallenge class
Currently just an unused stub though

### Improvements
#### minor UI things
- added 2 more workout sprites + videos
- changed ProgressView